create dhcpd.conf from template
This time I wrote a module specific feature in order to verify the
module does what it is supposed to do, i.e. create proper host entries.
File handling is not optimal, but enough for this showcase.

I use a different hostname to get a node that contains just this one
class and with the parameters I want it to have.

In the Then-step I again wrote the code I wish I had and implemented the
missing helper and matcher afterwards to keep my step implementations
simple.

This feature also helped design the dhcp::server class. Trying to inject
test data, using a parameterized class was the way to go. Without using
a test like this, I might have just hardcoded the file to use in the
class.

At this point I hardcoded the values in the template. I will refactor it
in the next step.